requirements.txt
streamlit
langchain
openai
faiss-cpu
python-docx
python-pptx
PyMuPDF
pandas
markdown
tiktoken
